Rename i_kubernetes_service file to _i_spark_job_service
